refactor: rearrange directory structure
- move public/index.html to root directory
- move files for the component to lib/
- rename src/lib/ to lib/utils/ to avoid lib/ under lib/
- move files to be served as a web app to web/
- include all files in tsconfig.json
